requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.4
django-crispy-forms==1.14.0
django-embed-video==1.4.8
django-jazzmin==2.6.0
django-phonenumber-field==7.0.1
django-tinymce==3.5.0
idna==3.4
phonenumbers==8.13.2
Pillow==9.3.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
